+++ src/core/http/plugins/response/__init__.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+
+"""
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+    Development Team: Stanislav WEB
+"""
+
+# noinspection PyPep8Naming
+from .collation import CollationResponsePlugin as collation
+# noinspection PyPep8Naming
+from .indexof import IndexofResponsePlugin as indexof
+# noinspection PyPep8Naming
+from .skipempty import SkipemptyResponsePlugin as skipempty
+# noinspection PyPep8Naming
+from .file import FileResponsePlugin as file